+++ src/competitions/competitions.js
@@ -6,6 +6,7 @@ const Go = require('./go/Go');
 const FourInARow = require('./four-in-a-row/FourInARow');
 const MsHackMan = require('./ms-hack-man/MsHackMan');
 const Golad = require('./golad/Golad');
+const TexasHoldem = require('./texas-holdem/TexasHoldem');
 
 module.exports = [
   new HackMan(),
@@ -16,4 +17,5 @@ module.exports = [
   new FourInARow(),
   new MsHackMan(),
   new Golad(),
+  new TexasHoldem(),
 ].sort((c1, c2) => c1.name.localeCompare(c2.name));